+++ mithril-aggregator/src/database/repository/cardano_transaction_repository.rs
@@ -313,6 +313,20 @@ impl TransactionsRetriever for CardanoTransactionRepository {
             .map(|record| record.into())
             .collect())
     }
+
+    async fn get_by_block_ranges(
+        &self,
+        block_ranges: Vec<BlockRange>,
+    ) -> StdResult<Vec<CardanoTransaction>> {
+        let provider = GetCardanoTransactionProvider::new(&self.connection);
+        let filters = provider.get_transaction_block_ranges_condition(block_ranges);
+        let transactions = provider.find(filters)?;
+
+        Ok(transactions
+            .into_iter()
+            .map(|record| record.into())
+            .collect())
+    }
 }
 
 #[async_trait]
@@ -489,7 +503,7 @@ mod tests {
         let connection = Arc::new(cardano_tx_db_connection().unwrap());
         let repository = CardanoTransactionRepository::new(connection);
 
-        // Build transactions with block numbers from 10 to 40 and immutable file numbers from 12 to 14
+        // Build transactions with block numbers from 20 to 40 and immutable file numbers from 12 to 14
         let cardano_transactions: Vec<CardanoTransactionRecord> = (20..=40)
             .map(|i| CardanoTransactionRecord {
                 transaction_hash: format!("tx-hash-{i}"),
@@ -674,6 +688,63 @@ mod tests {
             Some(last_block_range.end..(last_transaction_block_number + 1)),
             interval
         );
+    }
+
+    #[tokio::test]
+    async fn repository_get_transactions_by_block_ranges() {
+        let connection = Arc::new(cardano_tx_db_connection().unwrap());
+        let repository = CardanoTransactionRepository::new(connection);
+
+        let transactions = vec![
+            CardanoTransaction::new("tx-hash-1", 10, 50, "block-hash-1", 99),
+            CardanoTransaction::new("tx-hash-2", 11, 51, "block-hash-2", 100),
+            CardanoTransaction::new("tx-hash-3", 20, 52, "block-hash-3", 101),
+            CardanoTransaction::new("tx-hash-4", 31, 53, "block-hash-4", 102),
+            CardanoTransaction::new("tx-hash-5", 35, 54, "block-hash-5", 103),
+            CardanoTransaction::new("tx-hash-6", 46, 55, "block-hash-6", 104),
+        ];
+        repository
+            .create_transactions(transactions.clone())
+            .await
+            .unwrap();
+
+        {
+            let transaction_result = repository
+                .get_by_block_ranges(vec![BlockRange::from_block_number(100)])
+                .await
+                .unwrap();
+            assert_eq!(Vec::<CardanoTransaction>::new(), transaction_result);
+        }
+        {
+            let transaction_result = repository
+                .get_by_block_ranges(vec![BlockRange::from_block_number(0)])
+                .await
+                .unwrap();
+            assert_eq!(transactions[0..=1].to_vec(), transaction_result);
+        }
+        {
+            let transaction_result = repository
+                .get_by_block_ranges(vec![
+                    BlockRange::from_block_number(0),
+                    BlockRange::from_block_number(15),
+                ])
+                .await
+                .unwrap();
+            assert_eq!(transactions[0..=2].to_vec(), transaction_result);
+        }
+        {
+            let transaction_result = repository
+                .get_by_block_ranges(vec![
+                    BlockRange::from_block_number(0),
+                    BlockRange::from_block_number(30),
+                ])
+                .await
+                .unwrap();
+            assert_eq!(
+                [transactions[0..=1].to_vec(), transactions[3..=4].to_vec()].concat(),
+                transaction_result
+            );
+        }
     }
 
     #[tokio::test]